requirements.txt
selenium~=4.41.0
webdriver-manager~=4.0.2
requests
python-dotenv
PyPDF2
protobuf